requirements.txt
mysql-connector-python
fastapi
uvicorn
python-dotenv
pandas